Fix issue #987 - need to box value types if passed to ctor args as type object
If the value of a property of type int is used to populate a constructor argument of type object in deserialization then this is allowed by the constructor-matching code because it simply uses IsAssignableFrom as a check for compatability (and object IsAssignable from int) but when the IL is generated to pass an int value as an object, it has to be boxed in order for the code to be valid.

See notes at https://github.com/neuecc/MessagePack-CSharp/issues/987#issuecomment-683445396 that demonstrate how to use the reproduce case to generate a dll that PEVerify can be run against to show that it was invalid before this change.

I've added unit tests in this changeset for the fix but I suspect that the test runner would have difficulties executing the tests WITHOUT the fix as the ExecutionEngineError could tear down the process before the test runner could record it as a failure. However, if you were to use Debug Test on one of the tests that I've added but comment out the fix that I put into the DynamicObjectResolver then you would be able to see the ExecutionEngineError be caught by the debugger.